all tests are good

+++ app/controllers/subscription_controller.rb
@@ -12,14 +12,15 @@ class SubscriptionController < ApplicationController
     @user = User.find(params[:id])
     subs = current_user.subscriptions
     subs.each do |s|
-      if s.blogger == @user.id
-        s.destroy
-      end
+      s.destroy if s.blogger == @user.id
     end
     redirect_to user_path(@user)
   end
 
   def show
+  end
+
+  def index
   end
 
   def subscription_params
